// src/modules/settings/settings.model.d.ts
import mongoose, { Document } from "mongoose";
export interface ISetting extends Document {
    key: string;
    value: any;
    createdAt: Date;
    updatedAt: Date;
}
export declare const Setting: mongoose.Model<ISetting, {}, {}, {}, mongoose.Document<unknown, {}, ISetting, {}, mongoose.DefaultSchemaOptions> & ISetting & Required<{
    _id: mongoose.Types.ObjectId;
}> & {
    __v: number;
} & {
    id: string;
}, any, ISetting>;
//# sourceMappingURL=settings.model.d.ts.map